refactor: Update getFeaturedPosts function to accept a limit parameter
This commit modifies the getFeaturedPosts function in the posts.ts file to accept a limit parameter. Previously, the function had a hardcoded limit of 3, but now it can accept any number as the limit for fetching featured posts from the Ghost API. This change improves flexibility and allows for dynamic control of the number of featured posts to be retrieved.